Change the type of theDelay as bytes can only go to 255, and the user may want a longer delay time.

+++ Blank/Blank_NoDelay/Blank_NoDelay.ino
@@ -15,7 +15,7 @@
    User editable variables
 */
 // theDelay tracks how long to pause between colour changes
-byte theDelay = 100;
+int theDelay = 100;
 
 // theColour sets the colour that is used as the basis of the effect
 rgb_t theColour = BLUE;
